+++ src/components v2/Atoms/Forms/Autocomplete/Footer.vue
@@ -1,7 +1,7 @@
 <template>
   <div
-    v-if="showFooter"
-    class="ph-opacity-0 ph-transition ph-autocomplete__option--footer ph-flex ph-m-0 ph-px-5 ph-py-2 ph-text-sm ph-sticky ph-bottom-0 ph-z-1 ph-bg-white"
+    v-if="showFooter && addFooter"
+    class="ph-autocomplete__option--footer"
     :class="addFooter ? 'ph-opacity-100' : ''"
   >
     <slot name="list-footer"> </slot>
@@ -26,4 +26,21 @@ export default Vue.extend({
 });
 </script>
 
-<style></style>
+<style lang="scss">
+.ph-autocomplete__option--footer {
+  display: flex;
+  position: sticky;
+  bottom: 0;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  padding-left: 1.25rem;
+  padding-right: 1.25rem;
+  margin: 0;
+  background-color: #ffffff;
+  transition-property: background-color, border-color, color, fill, stroke,
+    opacity, box-shadow, transform;
+  font-size: 0.875rem;
+  line-height: 1.25rem;
+  opacity: 0;
+}
+</style>